refactor: rename folder "packages" to  "internal" to following golang project standard
