+++ internal/router/authentication.go
@@ -4,7 +4,7 @@ import (
 	"database/sql"
 
 	"github.com/go-chi/chi/v5"
-	"github.com/hebertzin/jwt-and-ratelimit-rest-api/packages/infra/factory"
+	"github.com/hebertzin/jwt-and-ratelimit-rest-api/internal/infra/factory"
 )
 
 func AuthenticationGroupRouter(r chi.Router, db *sql.DB) {
