+++ internal/services/users.go
@@ -3,11 +3,11 @@ package services
 import (
 	"context"
 
-	"github.com/hebertzin/jwt-and-ratelimit-rest-api/packages/domain"
-	"github.com/hebertzin/jwt-and-ratelimit-rest-api/packages/infra/repository"
-	"github.com/hebertzin/jwt-and-ratelimit-rest-api/packages/infra/security"
-	"github.com/hebertzin/jwt-and-ratelimit-rest-api/packages/utils"
-	"github.com/hebertzin/jwt-and-ratelimit-rest-api/packages/utils/validation"
+	"github.com/hebertzin/jwt-and-ratelimit-rest-api/internal/domain"
+	"github.com/hebertzin/jwt-and-ratelimit-rest-api/internal/infra/repository"
+	"github.com/hebertzin/jwt-and-ratelimit-rest-api/internal/infra/security"
+	"github.com/hebertzin/jwt-and-ratelimit-rest-api/internal/utils"
+	"github.com/hebertzin/jwt-and-ratelimit-rest-api/internal/utils/validation"
 )
 
 type UserService struct {
